requirements.txt
crewai
pandas
agentops
tabulate 
langchain 
streamlit
matplotlib
python-dotenv
langchain-core 
langchain-groq 
langchain-experimental
